pyproject.toml
[tool.poetry]
name = "iot_server"
version = "0.1.0"
description = ""
authors = ["mehonator <mehonator@yandex.ru>"]

[tool.poetry.dependencies]
python = "^3.9"
Django = "^4.0.1"
djangorestframework = "^3.13.1"
Markdown = "^3.3.6"
django-filter = "^21.1"
wakeonlan = "^2.1.0"

[tool.poetry.dev-dependencies]
flake8 = "^4.0.1"
mypy = "^0.931"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
